resume_ai_model/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
datasets
scikit-learn
pandas
numpy
pdfplumber
python-docx
pytesseract
Pillow
shap
spacy
nltk
pymupdf
pdfminer.six
easyocr
layoutparser
selenium
beautifulsoup4
paddleocr
opencv-python
